Move user auth tests
